requirements.txt
aiogram==3.19.0
openai==1.66.3
python-dotenv==1.0.0
aiofiles==23.2.1 